A bunch of JASTrack matching

+++ include/JSystem/JAudio/JASystem/JASRate.hpp
@@ -1,14 +1,12 @@
 #ifndef JASRATE_HPP
 #define JASRATE_HPP
 
-#include "types.h"
-
 #include <dolphin/types.h>
 
 namespace JASystem {
 namespace Kernel {
 	extern f32 gDacRate;
-	extern f32 gOutputRate;
+	extern u32 gOutputRate;
 	extern u32 gSubFrames;
 	extern u32 gDacSize;
 	extern u32 gFrameSamples;
