+++ src/JSystem/JAudio/JASystem/JASRate.cpp
@@ -1,9 +1,9 @@
-#include "JSystem/JAudio/JASystem/JASRate.hpp"
+#include <JSystem/JAudio/JASystem/JASRate.hpp>
 
 namespace JASystem {
 namespace Kernel {
 	f32 gDacRate      = 32028.5f;
-	f32 gOutputRate   = 0.0f;
+	u32 gOutputRate   = 0;
 	u32 gSubFrames    = 0x7;
 	u32 gFrameSamples = 0x230;
 	u32 gDacSize      = 0x460;
